+++ code/api/src/modules/product/resolvers.js
@@ -43,43 +43,53 @@ export async function getRelated(parentValue, { productId }) {
 }
 
 // Create product
-export async function create(parentValue, { name, slug, description, type, gender, image }) {
-
-  return await models.Product.create({
-    name,
-    slug,
-    description,
-    type,
-    gender,
-    image
-  })
-}
-
-// Update product
-export async function update(parentValue, { id, name, slug, description, type, gender, image }) {
-
-  return await models.Product.update(
-    {
+export async function create(parentValue, { name, slug, description, type, gender, image }, { auth }) {
+  if(auth.user && auth.user.role === params.user.roles.admin) {
+    return await models.Product.create({
       name,
       slug,
       description,
       type,
       gender,
       image
-    },
-    { where: { id } }
-  )
+    })
+  } else {
+    throw new Error('Operation denied.')
+  }
+}
+
+// Update product
+export async function update(parentValue, { id, name, slug, description, type, gender, image }, { auth }) {
+  if(auth.user && auth.user.role === params.user.roles.admin) {
+    return await models.Product.update(
+      {
+        name,
+        slug,
+        description,
+        type,
+        gender,
+        image
+      },
+      { where: { id } }
+    )
+  } else {
+    throw new Error('Operation denied.')
+  }
 }
 
 // Delete product
-export async function remove(parentValue, { id }) {
-  const product = await models.Product.findOne({ where: { id } })
+export async function remove(parentValue, { id }, { auth }) {
+  if(auth.user && auth.user.role === params.user.roles.admin) {
+    const product = await models.Product.findOne({where: {id}})
 
-  if (!product) {
-    // Product does not exists
-    throw new Error('The product does not exists.')
+    if (!product) {
+      // Product does not exists
+      throw new Error('The product does not exists.')
+    } else {
+      return await models.Product.destroy({where: {id}})
+    }
   } else {
-    return await models.Product.destroy({ where: { id } })
+    throw new Error('Operation denied.')
   }
 }
